Instances exec: send first frame even without data
We used to wait for some data to be intput on stdin before sending the
very first frame of the handshake. This is not needed for interactive
sessions (it adds a small overhead, too). For non-interactive sessions,
it requires the user to input text that he does not care about