Replace renderer.Format<obj>ID with a generic function
The function Format<obj>ID was called to display the UUID of the object
in the short form, or the long form if --full is provided. This function
was using the idmapper, which lists all the <obj> to return the only
object matching the id. This is unnecessary, and instead, we simply
return the first 8 digits of the uuid to display the short form.