Error rendering: always display the orig error when Orig is set
I don't know why, but in the case `koyeb` is called with `-f xxx`, where
xxx doesn't have an extension, a CLIError is created with Orig set to a
non-nil value, but the original error was not displayed.

With the custom function isNil, the error is correctly displayed only if
the field Orig is set.