Add --git-sha to koyeb service update, closes #222
Warning: if the service is deployed with a specific git sha (which is
the case if the deployment has been triggered by the "git push"
webhook), `koyeb service update` without `--git-sha` will no longer
deploy the latest commit, but the commit set in the configuration. To
deploy the latest commit, use `koyeb service update --git-sha ''`.